verificaçoes primárias

+++ src/services/transactionsService.js
@@ -8,11 +8,21 @@ export async function getTransactionsService() {
 }
 
 export async function createTransactionService(data) {
+    if (data.amount <= 0) {
+        throw new Error("O valor da transação deve ser maior que zero")
+    }
+
     const createTransaction = await transactionsRepository.create(data)
     return createTransaction
 }
 
 export async function deleteTransactionService(id){
+    const transaction = await transactionsRepository.findById(id)
+
+    if (!transaction) {
+        throw new Error("Transação não encontrada")
+    }
+
     const deleteTransaction = await transactionsRepository.delete(id)
     return deleteTransaction
 }
@@ -20,4 +30,9 @@ export async function deleteTransactionService(id){
 export async function updateTransactionService(id, data) {
     const updateTransaction = await transactionsRepository.update(id, data)
     return updateTransaction
+}
+
+export async function findTransactionByIdService(id) {
+    const transactionFound = await transactionsRepository.findById(id)
+    return transactionFound
 }
